+++ android/app/src/main/java/com/zzang/chongdae/data/mapper/MemberMapper.kt
@@ -1,7 +1,7 @@
 package com.zzang.chongdae.data.mapper
 
 import com.zzang.chongdae.data.remote.dto.response.MemberResponse
-import com.zzang.chongdae.domain.mo.Member
+import com.zzang.chongdae.domain.model.Member
 
 fun MemberResponse.toDomain(): Member {
     return Member(
