+++ android/app/src/main/java/com/zzang/chongdae/domain/model/Member.kt
@@ -1,4 +1,4 @@
-package com.zzang.chongdae.domain.mo
+package com.zzang.chongdae.domain.model
 
 data class Member(
     val memberId: Long,
